requirements.txt
python-docx
transformers
torch
torchvision
torchaudio